// src/app/features/lms/ind/components/claims/models/claims.ts
export interface ClmCvtReserveDTO {
  pol_code: number;
  cvt_code: number;
  cvt_sht_desc: string;
  cvt_desc: string;
  pcvt_sa: number;
  cnoc_reserved_amt: number;
  value: number;
  cnot_code: number;
  clm_no: string;
}

export interface ClaimDTO {
  clm_no: string;
  prp_code: number;
  pol_code: number;
  pol_policy_no: string;
  client: string;
  pol_prod_sht_desc: string;
  brn_name: string;
  pol_effective_date: string;
  pol_maturity_date: string;
  cnot_code: number;
  cnot_death_date: string;
  cnot_disability_date: string | null;
  cnot_notification_date: string;
  caus_desc: string;
  ddc_desc: string;
  clm_admitted: string;
  clm_air_claim_type: string | null;
  clm_status: string;
  clm_bonus_paid: string | null;
  clm_date_death_accident: string;
  clm_date_claim_reported: string;
  clm_retention_pct: number;
  clm_amt_claimed: number | null;
  clm_paid_amt: number | null;
  clm_amt_to_pay: number | null;
  clm_os_premium: number;
  clm_os_loan: number;
  clm_pol_sa: number;
  clm_indmty_wks: number | null;
  clm_hosp_days: number | null;
  clm_payee: string | null;
  clm_paid_matrty_amt: number | null;
  clm_retetion_amt: number | null;
  clm_encash_clm_instls: string;
  clm_adv_charge: number | null;
  clm_caus_type: string;
  tot_loss_perc: number | null;
  tot_bonus: number | null;
  clm_reserve_amt: number;
  clm_chkoff_comm_amt: number | null;
  clm_prem_waiver: string;
  endr_code: number;
  net_commutatn: number | null;
  commutate: string;
  injectn_prem: number;
  commutatn_perc: number;
  remarks: string | null;
  caus_code: number;
  terminalbonus: number | null;
  sa_topay_instlmnts: number | null;
  pol_prod_type: string;
  pol_prod_code: number;
  endr_ann_repay_freq: string | null;
  clm_refund_amt: number | null;
  pol_grp_life_rider: string;
  display_life_riders: string;
  pol_paid_to_date: string;
  clm_dis_end_date: string | null;
  clm_death_location: string;
  clm_no_claim_disc_amt: number;
  clm_but_pay_amount: number;
  clm_but_pay_reason: string | null;
  clm_disinvested_amnt: number | null;
  clm_payee_address: string | null;
  clm_payee_tel: string | null;
  clm_overide_inc_prd: string;
  clm_unproc_bal_amt: number;
  clm_pay_lump_sum: number | null;
  prod_invest_all_prem: number | null;
  tsd_contract_status: string | null;
  tsd_paying_status: string | null;
  tss_desc: string | null;
  pol_status: string;
  assigned: string;
  clm_but_pay_checked: string | null;
  tot_opr_amt: number;
  endr_tot_sa: number;
  premium: number;
  pol_freq_of_payment: string;
  clm_exgratia_checked: string;
  clm_exgratia_reason: string | null;
  fund_value: number;
  clm_repudiation_remarks: string | null;
  prod_allow_dep_prem_inv: string;
  prod_inv_rider_allowed: string;
  clm_dep_prem_inv: string;
  endr_anb: number;
  endr_dob: string;
  clm_new_age: number | null;
  clm_reinstmnt_interest: number | null;
  claim_life: string;
  clm_processing_brn_code: number | null;
  processing_brn_name: string | null;
  clm_pod_unique_code: number;
  clm_pod_code: number;
  clm_rider_surr_val: number;
  clm_charge: number;
  clm_prepared_by: string;
  clm_cvt_reserves: ClmCvtReserveDTO[];
  clm_pay_currency: string;
  clm_pay_installments: number | null;
  ddc_code: number | null;
}