// src/app/features/lms/ind/components/claims/models/policy-claim-intiation.ts
export interface PolicyClaimIntiation {
  pol_code: number;
  clm_date_death_accident: string;
  place_of_incident: string;
  clm_exgratia_checked: string;
  caus_code: number;
  caus_sht_desc: string;
  csc_code: number;
  ddc_code: number;
  ddc_desc: string;
  place_of_death: string;
  caus_type: string;
  delete_existing_trans: string;
  clm_pod_code: number;
  pol_policy_no: string;
  prp_code: number;
}